refactor(decode): Scalar monolith + drop x86 macro + one-time dispatch
Adds `seq_decoder_scalar.rs` as the Scalar-tier monolith (same shape
as the AVX2 / BMI2 / VBMI2 monoliths, pinned to `ScalarKernel`, no
target_feature). Dispatcher in
`sequence_section_decoder::decode_and_execute_sequences` now routes
the Scalar arm into the new monolith too.

Deletes `seq_decoder_x86_kernel.rs` entirely — the
`define_x86_seq_decoder_tier!` macro is no longer referenced by any
tier file. Each tier (Scalar / BMI2 / AVX2 / VBMI2) is now a fully
self-contained function with the whole decode + execute pipeline
inlined in one place, selected ONCE per call via cached
`detect_cpu_kernel`.

aarch64 NEON / SVE still go through the K-generic
`decode_and_execute_sequences_impl` shared body until their own
monoliths land in a follow-up.
