+++ zstd/src/decoding/seq_decoder_scalar.rs
@@ -1,0 +1,292 @@
+//! Scalar-tier monolithic sequence-section decoder.
+//!
+//! One self-contained function with the entire decode + execute
+//! pipeline inlined. BitReader pinned to `ScalarKernel` (no PEXT, no
+//! SIMD wildcopy — pure portable arithmetic + libc memmove fallback in
+//! the match-copy path). Used on any non-x86_64 target without the
+//! corresponding aarch64 monolith available, and on x86_64 when
+//! `detect_cpu_kernel()` falls through to Scalar (extremely old CPUs).
+
+use super::buffer_backend::BufferBackend;
+use super::decode_buffer::DecodeBuffer;
+use super::scratch::FSEScratch;
+use super::sequence_section_decoder::{
+    ADVANCE, ADVANCE_MASK, ExecSeq, compute_use_long_pipeline, decode_sequences_with_rle,
+    execute_one_sequence_pipelined, maybe_update_fse_tables,
+};
+use crate::bit_io::BitReaderReversed;
+use crate::blocks::sequence_section::{MAX_OFFSET_CODE, Sequence, SequencesHeader};
+use crate::common::MAX_BLOCK_SIZE;
+use crate::cpu_kernel::ScalarKernel;
+use crate::decoding::errors::{DecodeSequenceError, DecompressBlockError, ExecuteSequencesError};
+use crate::decoding::sequence_execution::{do_offset_history, execute_sequences_fields};
+use crate::fse::SeqFSEDecoder;
+use alloc::vec::Vec;
+
+/// Scalar-tier monolithic decode + execute.
+pub(crate) fn decode_and_execute_sequences_scalar<B: BufferBackend>(
+    section: &SequencesHeader,
+    source: &[u8],
+    fse: &mut FSEScratch,
+    buffer: &mut DecodeBuffer<B>,
+    offset_hist: &mut [u32; 3],
+    literals_buffer: &[u8],
+    rle_fallback_sequences: &mut Vec<Sequence>,
+) -> Result<(), DecompressBlockError> {
+    rle_fallback_sequences.clear();
+
+    let ddict_is_cold = fse.ddict_is_cold;
+    fse.ddict_is_cold = false;
+
+    let bytes_read = maybe_update_fse_tables(section, source, fse)?;
+    let bit_stream = &source[bytes_read..];
+    let mut br = BitReaderReversed::<ScalarKernel>::new(bit_stream);
+
+    let mut skipped_bits = 0;
+    loop {
+        let val = br.get_bits(1);
+        skipped_bits += 1;
+        if val == 1 || skipped_bits > 8 {
+            break;
+        }
+    }
+    if skipped_bits > 8 {
+        return Err(DecodeSequenceError::ExtraPadding { skipped_bits }.into());
+    }
+
+    if fse.ll_rle.is_some() || fse.ml_rle.is_some() || fse.of_rle.is_some() {
+        decode_sequences_with_rle(section, &mut br, fse, rle_fallback_sequences)?;
+        execute_sequences_fields(buffer, literals_buffer, offset_hist, rle_fallback_sequences)?;
+        return Ok(());
+    }
+
+    let mut ll_dec = SeqFSEDecoder::new(&fse.literal_lengths);
+    let mut ml_dec = SeqFSEDecoder::new(&fse.match_lengths);
+    let mut of_dec = SeqFSEDecoder::new(&fse.offsets);
+
+    ll_dec
+        .init_state(&mut br)
+        .map_err(DecodeSequenceError::from)?;
+    of_dec
+        .init_state(&mut br)
+        .map_err(DecodeSequenceError::from)?;
+    ml_dec
+        .init_state(&mut br)
+        .map_err(DecodeSequenceError::from)?;
+
+    let max_update_bits = fse.literal_lengths.accuracy_log
+        + fse.match_lengths.accuracy_log
+        + fse.offsets.accuracy_log;
+    debug_assert!(
+        max_update_bits <= 56,
+        "sequence section update bits exceed 56-bit budget"
+    );
+
+    buffer.reserve(MAX_BLOCK_SIZE as usize);
+    let old_buffer_size = buffer.len();
+    let literals_buffer_len = literals_buffer.len();
+    let mut lit_cur: usize = 0;
+    let mut seq_sum: u32 = 0;
+
+    let buffer_checkpoint = buffer.checkpoint();
+    let saved_offset_hist = *offset_hist;
+    let num_sequences = section.num_sequences as usize;
+
+    let total_history = buffer.window_size + buffer.dict_content.len();
+    let use_long_pipeline = compute_use_long_pipeline(
+        num_sequences,
+        ddict_is_cold,
+        total_history,
+        fse.offsets_long_share,
+    );
+
+    if use_long_pipeline {
+        let mut prefetch_pos: usize = old_buffer_size;
+        let mut shadow_hist: [u32; 3] = *offset_hist;
+        let mut ring: [ExecSeq; ADVANCE] = [ExecSeq {
+            ll: 0,
+            ml: 0,
+            actual_offset: 0,
+        }; ADVANCE];
+
+        for slot in ring.iter_mut() {
+            let seq = decode_one_scalar(&mut ll_dec, &mut ml_dec, &mut of_dec, &mut br);
+            let actual_offset = do_offset_history(seq.of, seq.ll, &mut shadow_hist);
+            let match_start = prefetch_pos.wrapping_add(seq.ll as usize);
+            let source_idx = match_start.wrapping_sub(actual_offset as usize);
+            buffer.prefetch_lookahead_match_source(source_idx);
+            prefetch_pos = match_start.wrapping_add(seq.ml as usize);
+            *slot = ExecSeq {
+                ll: seq.ll,
+                ml: seq.ml,
+                actual_offset,
+            };
+            br.ensure_bits(max_update_bits);
+            ll_dec.update_state_fast(&mut br);
+            ml_dec.update_state_fast(&mut br);
+            of_dec.update_state_fast(&mut br);
+        }
+
+        let mut pipeline_err: Option<DecompressBlockError> = None;
+        for i in ADVANCE..num_sequences {
+            let seq = decode_one_scalar(&mut ll_dec, &mut ml_dec, &mut of_dec, &mut br);
+            let actual_offset = do_offset_history(seq.of, seq.ll, &mut shadow_hist);
+            let match_start = prefetch_pos.wrapping_add(seq.ll as usize);
+            let source_idx = match_start.wrapping_sub(actual_offset as usize);
+            buffer.prefetch_lookahead_match_source(source_idx);
+            prefetch_pos = match_start.wrapping_add(seq.ml as usize);
+
+            let slot = i & ADVANCE_MASK;
+            let exec_seq = ring[slot];
+            ring[slot] = ExecSeq {
+                ll: seq.ll,
+                ml: seq.ml,
+                actual_offset,
+            };
+
+            if let Err(e) = execute_one_sequence_pipelined(
+                buffer,
+                literals_buffer,
+                &mut lit_cur,
+                literals_buffer_len,
+                Sequence {
+                    ll: exec_seq.ll,
+                    ml: exec_seq.ml,
+                    of: 0,
+                },
+                exec_seq.actual_offset,
+            ) {
+                pipeline_err = Some(e);
+                break;
+            }
+            seq_sum = seq_sum.wrapping_add(exec_seq.ll).wrapping_add(exec_seq.ml);
+
+            if i + 1 < num_sequences {
+                br.ensure_bits(max_update_bits);
+                ll_dec.update_state_fast(&mut br);
+                ml_dec.update_state_fast(&mut br);
+                of_dec.update_state_fast(&mut br);
+            }
+        }
+
+        if pipeline_err.is_none() {
+            for k in 0..ADVANCE {
+                let slot = (num_sequences + k) & ADVANCE_MASK;
+                let exec_seq = ring[slot];
+                if let Err(e) = execute_one_sequence_pipelined(
+                    buffer,
+                    literals_buffer,
+                    &mut lit_cur,
+                    literals_buffer_len,
+                    Sequence {
+                        ll: exec_seq.ll,
+                        ml: exec_seq.ml,
+                        of: 0,
+                    },
+                    exec_seq.actual_offset,
+                ) {
+                    pipeline_err = Some(e);
+                    break;
+                }
+                seq_sum = seq_sum.wrapping_add(exec_seq.ll).wrapping_add(exec_seq.ml);
+            }
+        }
+
+        if let Some(e) = pipeline_err {
+            if buffer.try_restore_checkpoint(buffer_checkpoint) {
+                *offset_hist = saved_offset_hist;
+            }
+            return Err(e);
+        }
+        *offset_hist = shadow_hist;
+    } else {
+        let mut shadow_hist = *offset_hist;
+        let mut fallback_err: Option<DecompressBlockError> = None;
+        for i in 0..num_sequences {
+            let seq = decode_one_scalar(&mut ll_dec, &mut ml_dec, &mut of_dec, &mut br);
+            let resolved_offset = do_offset_history(seq.of, seq.ll, &mut shadow_hist);
+            if let Err(e) = execute_one_sequence_pipelined(
+                buffer,
+                literals_buffer,
+                &mut lit_cur,
+                literals_buffer_len,
+                seq,
+                resolved_offset,
+            ) {
+                fallback_err = Some(e);
+                break;
+            }
+            seq_sum = seq_sum.wrapping_add(seq.ll).wrapping_add(seq.ml);
+
+            if i + 1 < num_sequences {
+                br.ensure_bits(max_update_bits);
+                ll_dec.update_state_fast(&mut br);
+                ml_dec.update_state_fast(&mut br);
+                of_dec.update_state_fast(&mut br);
+            }
+        }
+        if let Some(e) = fallback_err {
+            let _ = buffer.try_restore_checkpoint(buffer_checkpoint);
+            return Err(e);
+        }
+        *offset_hist = shadow_hist;
+    }
+
+    let remaining = br.bits_remaining();
+    if remaining != 0 {
+        if buffer.try_restore_checkpoint(buffer_checkpoint) {
+            *offset_hist = saved_offset_hist;
+        }
+        if remaining < 0 {
+            return Err(DecodeSequenceError::NotEnoughBytesForNumSequences.into());
+        }
+        return Err(DecodeSequenceError::ExtraBits {
+            bits_remaining: remaining,
+        }
+        .into());
+    }
+
+    if lit_cur < literals_buffer_len {
+        let rest = &literals_buffer[lit_cur..];
+        buffer.try_push(rest).map_err(ExecuteSequencesError::from)?;
+        seq_sum = seq_sum.wrapping_add(rest.len() as u32);
+    }
+
+    let diff = buffer.len() - old_buffer_size;
+    debug_assert_eq!(
+        seq_sum as usize, diff,
+        "seq_sum {seq_sum} != buffer growth {diff}"
+    );
+    Ok(())
+}
+
+#[inline]
+fn decode_one_scalar(
+    ll_dec: &mut SeqFSEDecoder<'_>,
+    ml_dec: &mut SeqFSEDecoder<'_>,
+    of_dec: &mut SeqFSEDecoder<'_>,
+    br: &mut BitReaderReversed<'_, ScalarKernel>,
+) -> Sequence {
+    let ll_state = ll_dec.state;
+    let ml_state = ml_dec.state;
+    let of_state = of_dec.state;
+
+    let ll_value = ll_state.base_value;
+    let ll_num_bits = ll_state.num_additional_bits;
+    let ml_value = ml_state.base_value;
+    let ml_num_bits = ml_state.num_additional_bits;
+    let of_num_bits = of_state.num_additional_bits;
+    let of_base = of_state.base_value;
+
+    debug_assert!(of_num_bits <= MAX_OFFSET_CODE);
+
+    let (obits, ml_add, ll_add) = br.get_bits_triple(of_num_bits, ml_num_bits, ll_num_bits);
+    let offset = obits as u32 + of_base;
+    debug_assert_ne!(offset, 0);
+
+    Sequence {
+        ll: ll_value + ll_add as u32,
+        ml: ml_value + ml_add as u32,
+        of: offset,
+    }
+}
